Added inverted pendulum and a file for optimizing
Added subdirectory nlp_problems. This directory contains nonlinear programming problems. Right now it just has the inverted pendulum problem, and an abstract class for nla problems. The abstract class shows the necessary interface, such as functions for the gradient, Hessian, etc.

Also added a file for testing the optimization of a file.